third_party/blink/web_tests/external/wpt/custom-elements/revamped-scoped-registry/Document-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on a document of a disconnected iframe should return contentWindow.customElements
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
